Adds file storage guidelines to documentation
Were trying to use this for a hackathon just now. Took us a while to figure out what was wrong and only by inference. Don't have time right now to do much (still hacking) so if files are not required to be in the `public/` folder I apologize but that is the only thing that got this library working for us 👍